Mais testes e imagens

+++ Projeto/StarsDataset.py
@@ -2,6 +2,7 @@ import pandas as pd
 from sklearn.model_selection import train_test_split
 from sklearn.preprocessing import StandardScaler, LabelEncoder
 
+import h2o
 
 FILENAME = "Stars.csv"
 TRAIN_TO_TEST_RATIO = 3/4
@@ -46,7 +47,7 @@ class Dataset(object):
 
 class DatasetPanda(Dataset):
     def __init__(self, target="Classification"):
-        self.original = pd.read_csv(FILENAME)
+        self.read_database()
         self.setup_lists()
         self.normalize_color()
         self.define_target(target)
@@ -55,6 +56,11 @@ class DatasetPanda(Dataset):
         self.set_target()
         self.split()
         self.scale()
+        self.categorize(target)        
+        
+    def read_database(self):        
+        self.original = pd.read_csv(FILENAME)     
+        self.data = self.original.copy()
     
     # Lists end up Class variables instead of Instance variables if not done here
     def setup_lists(self):
@@ -79,55 +85,62 @@ class DatasetPanda(Dataset):
             "Type",
         ]        
         
-    # Color is really unorganized, this is a bit unfair but helps
+    # Color is too unorganized, this changes similar colors to their base color
     def normalize_color(self):
         col = "Color"        
-        self.original[col] = self.original[col].str.lower()
-        self.original[col] = self.original[col].str.replace(" ", "-")
-        self.original[col] = self.original[col].str.replace("whitish", "white")
-        self.original[col] = self.original[col].str.replace("ish", "")
-        self.original[col] = self.original[col].str.replace("orange-red", "orange")
-        self.original[col] = self.original[col].str.replace("pale-yellow-orange", "orange")
-        self.original[col] = self.original[col].str.replace("white-yellow", "yellow")
+        self.data[col] = self.data[col].str.lower()
+        self.data[col] = self.data[col].str.replace(" ", "-")
+        self.data[col] = self.data[col].str.replace("whitish", "white")
+        self.data[col] = self.data[col].str.replace("ish", "")
+        self.data[col] = self.data[col].str.replace("orange-red", "orange")
+        self.data[col] = self.data[col].str.replace("pale-yellow-orange", "yellow-white")
+        self.data[col] = self.data[col].str.replace("white-yellow", "yellow-white")
     
     # Sets the index and column name
     def define_target(self, target_label):
-        self.target_index = self.label_to_index(target_label)
-        self.target_column = self.label_to_column(target_label)
+        self.target_index = None
+        self.target_column = None
+        if target_label:
+            self.target_index = self.label_to_index(target_label)
+            self.target_column = self.label_to_column(target_label)
     
     # Gets the names before text is transformed into numbers
     def set_target_labels(self):
+        if not self.target_column:
+            return
         if self.target_column == "Type":
             self.target_classes = CLASSIFICATION
             return
-        target_values = self.original[self.target_column]
+        target_values = self.data[self.target_column]
         self.target_classes = list(set(target_values))
     
     # Changes text into numbers
     def flatten_data(self):
         label_encoder = LabelEncoder()
         text_columns = ["Color", "Spectral_Class"]
-        flattened_text = self.original[text_columns].values.flatten()
+        flattened_text = self.data[text_columns].values.flatten()
         label_encoder.fit(flattened_text)
         flattened_columns = label_encoder.fit_transform
-        self.original[text_columns] = self.original[text_columns].apply(flattened_columns)
+        self.data[text_columns] = self.data[text_columns].apply(flattened_columns)
     
     # Gets the target values and drops it from the data
     def set_target(self):
-        self.target = self.original[self.target_column]
-        self.drop_column(self.target_column)
+        self.target = None
+        if self.target_column:
+            self.target = self.data[self.target_column]
+            self.drop_column(self.target_column)
     
     # Removes from the data, then the lists
     def drop_labels(self, labels=None):
         if labels:
             for label in labels:
                 column = self.label_to_column(label)
-                self.original = self.original.drop(columns=column)
+                self.data = self.data.drop(columns=column)
             self.drop_labels_from_lists(labels)
     
     # Same as above, but "private"
     def drop_column(self, column):
-        self.original = self.original.drop(columns=column)
+        self.data = self.data.drop(columns=column)
         self.LABELS.pop(self.column_to_index(column))
         self.COLUMNS.pop(self.column_to_index(column))
     
@@ -139,8 +152,12 @@ class DatasetPanda(Dataset):
         
     # Splits into test and training in a balanced manner
     def split(self):
+        if self.target_column == None:
+            self.train = None
+            self.test = None
+            return
         self.train, self.test, self.train_classes, self.test_classes = train_test_split(
-            self.original, 
+            self.data, 
             self.target, 
             train_size=TRAIN_TO_TEST_RATIO, 
             stratify=self.target)        
@@ -148,10 +165,31 @@ class DatasetPanda(Dataset):
     # "Standardizes features by removing the mean and scaling to unit variance"
     # Helps on simpler comparations
     def scale(self):
+        if self.target_column == None:
+            return
         scaler = StandardScaler()
         scaler.fit(self.train)
         self.scaled_train = scaler.transform(self.train)
         self.scaled_test = scaler.transform(self.test)
-        
+    
+    def categorize(self, target):
+        self.uncategorized = self.data.copy()
+        if target != "Color":
+            self.data["Color"] = pd.Categorical(self.data["Color"])
+        if target != "SMASS":
+            self.data["Spectral_Class"] = pd.Categorical(self.data["Spectral_Class"])
+        if target != "Classification":
+            self.data["Type"] = pd.Categorical(self.data["Type"])
         
 
+# Ended up not being used
+class DatasetH2O(DatasetPanda):
+    def __init__(self):
+        super().__init__("")
+        h2o.init(ip="127.0.0.1", port="8080")
+        self.data_hex = h2o.H2OFrame(self.data)
+        # self.data_hex["Type"] = h2o.H2OFrame(self.target).asFactor()
+        self.data_train_hex, self.data_test_hex = self.data_hex.split_frame(
+            ratios=[TRAIN_TO_TEST_RATIO], 
+            seed=999)
+        print(self.data_train_hex["Type"].as_data_frame().value_counts())
